Create packageManager
